Change OxGame to OxRound and add new OxGame class
