+++ src/main/java/akademia/ox/states/DrawState.java
@@ -42,7 +42,7 @@ public class DrawState implements GameState {
     }
 
     @Override
-    public OxGame showGame() {
+    public OxRound showGame() {
         return null;
     }
 }
