+++ src/main/java/akademia/ox/states/TerminateState.java
@@ -34,7 +34,7 @@ public class TerminateState implements GameState {
     }
 
     @Override
-    public OxGame showGame() {
+    public OxRound showGame() {
         return null;
     }
 }
