+++ src/main/java/akademia/ox/states/VictoryState.java
@@ -40,7 +40,7 @@ public class VictoryState implements GameState {
     }
 
     @Override
-    public OxGame showGame() {
+    public OxRound showGame() {
         return null;
     }
 }
