+++ src/test/java/akademia/ox/StateMachineWithPlayersTests.java
@@ -12,7 +12,7 @@ public class StateMachineWithPlayersTests {
     private Player playerTwo = new Player("p2", "O");
     private BoardVisualizer bv = new BoardVisualizer();
     private VictoryChecker vc = new VictoryChecker();
-    private OxGame game;
+    private OxRound game;
 
 
     private void setPlayers() throws IncorrectPlayerException, TooManyPlayersException {
@@ -22,7 +22,7 @@ public class StateMachineWithPlayersTests {
     }
 
     private void setGame() {
-        game = OxGame.createStandardGame(bv, vc);
+        game = OxRound.createStandardGame(bv, vc);
     }
 
     @Test
